move page reader types to page_reader.go
